+++ src/main/java/com/ydo4ki/bullsandcows/Main.java
@@ -1,6 +1,6 @@
 package com.ydo4ki.bullsandcows;
 
-import java.util.Scanner;
+import java.util.*;
 import java.util.concurrent.ThreadLocalRandom;
 
 public class Main {
@@ -11,14 +11,26 @@ public class Main {
 		Scanner scanner = new Scanner(System.in);
 		
 		System.out.println("Welcome! Enter amount of digits to guess: ");
-		int digits = scanner.nextInt();
+		int digits;
+		while(true) try {
+			String line = scanner.nextLine();
+			digits = Integer.parseInt(line);
+			if (digits < 1 || digits > 9) {
+				System.out.println("Amount must be between 1 and 9");
+				continue;
+			}
+			break;
+		} catch (NumberFormatException e) {
+			System.out.println("Invalid input :(");
+			continue;
+		}
 		String number = generateNumber(digits);
 		if (DEBUG) System.out.println(number);
 		
 		System.out.println(
 				"Number guessed, now try to guess it.\n" +
-				"The output will indicate how far are you from the number:\n" +
-				"bull means that your guess contains a correct digit in a right place, cow means that your guess contains a correct digit in a wrong place"
+						"The output will indicate how far are you from the number:\n" +
+						"bull means that your guess contains a correct digit in a right place, cow means that your guess contains a correct digit in a wrong place"
 		);
 		while(true) {
 			String input;
@@ -44,9 +56,12 @@ public class Main {
 	}
 	
 	private static String generateNumber(int digits) {
+		List<Character> possibleDigits = new ArrayList<>();
+		Collections.addAll(possibleDigits, '1', '2', '3', '4', '5', '6', '7', '8', '9');
+		
 		char[] symbols = new char[digits];
 		for (int i = 0; i < symbols.length; i++) {
-			symbols[i] = (char)ThreadLocalRandom.current().nextInt('1','9'+1);
+			symbols[i] = possibleDigits.remove(ThreadLocalRandom.current().nextInt(0, possibleDigits.size()));
 		}
 		return String.valueOf(symbols);
 	}